warning fixed
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@15559 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/tad/ipstack/tad_tcp_layer.c
@@ -272,7 +272,7 @@ tcp_gen_bin_cb(csap_p csap_descr, int layer, const asn_value *tmpl_pdu,
             switch (length_)                                            \
             {                                                           \
                 case 1:                                                 \
-                    *((uint8_t *)p) = (def_val_);                       \
+                    *((uint8_t *)p) = (uint8_t)(def_val_);              \
                     break;                                              \
                                                                         \
                 case 2:                                                 \
